Modularization from the home component
