+++ src/app/components/tools/img-techs/img-techs.component.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+
+<section class="tech-stack">
+  <h2 class="tech-title">Tecnologías utilizadas</h2>
+  <div class="tech-logos">
+    <img src="assets/images/technologies/Angular.png" alt="Angular" title="Framework de desarrollo web">
+    <img src="assets/images/technologies/NetFreamwork.png" alt=".NET Framework" title="Api RESTful para manejo de JWT y lógica de negocio">
+    <img src="assets/images/technologies/SqlServer.png" alt="SQL Server" title="Database relacional para el sistema">
+    <img src="assets/images/technologies/firebase.png" alt="Firebase" title="Firebase: Realtime DB y hosting para imágenes y errores">
+  </div>
+</section>
