+++ src/app/components/tools/img-techs/img-techs.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-img-techs',
+  standalone: false,
+  templateUrl: './img-techs.component.html',
+  styleUrl: './img-techs.component.css'
+})
+export class ImgTechsComponent {
+
+}
